refactor: started moving the chat inspector to an independent view
