+++ Shared/Views/ChatPicker.swift
@@ -1,5 +1,5 @@
 //
-//  ChatPickerView.swift
+//  ChatPicker.swift
 //  Moc
 //
 //  Created by Егор Яковенко on 02.06.2022.
@@ -7,8 +7,8 @@
 
 import SwiftUI
 
-// TODO: Implement ChatPickerView
-struct ChatPickerView: View {
+// TODO: Implement ChatPicker
+struct ChatPicker: View {
     var body: some View {
         VStack {
             SearchField()
@@ -22,6 +22,6 @@ struct ChatPickerView: View {
 
 struct ChatPickerView_Previews: PreviewProvider {
     static var previews: some View {
-        ChatPickerView()
+        ChatPicker()
     }
 }
